SAA-501: Rename e2e tests
