+++ integration_tests/pages/appointments/create/categoryPage.ts
